core-android-mock module renamed to environment-android-mock
