+++ advertiser-android-mock/build.gradle.kts
@@ -48,7 +48,7 @@ nordicNexusPublishing {
 
 dependencies {
     api(project(":advertiser-core-android"))
-    api(project(":core-android-mock"))
+    api(project(":environment-android-mock"))
 
     implementation(libs.slf4j)
 }
