+++ advertiser-android-mock/src/main/java/no/nordicsemi/kotlin/ble/advertiser/android/mock/BluetoothLeAdvertiserFactory.kt
@@ -33,7 +33,7 @@ package no.nordicsemi.kotlin.ble.advertiser.android.mock
 
 import no.nordicsemi.kotlin.ble.advertiser.android.BluetoothLeAdvertiser
 import no.nordicsemi.kotlin.ble.advertiser.android.mock.internal.MockBluetoothLeAdvertiser
-import no.nordicsemi.kotlin.ble.android.mock.MockAndroidEnvironment
+import no.nordicsemi.kotlin.ble.environment.android.mock.MockAndroidEnvironment
 
 /**
  * Creates an instance of a mock [BluetoothLeAdvertiser] for Android.
