+++ client-android-mock/src/main/java/no/nordicsemi/kotlin/ble/client/android/mock/CentralManagerFactory.kt
@@ -34,11 +34,11 @@
 package no.nordicsemi.kotlin.ble.client.android.mock
 
 import kotlinx.coroutines.CoroutineScope
-import no.nordicsemi.kotlin.ble.android.mock.LatestApi
-import no.nordicsemi.kotlin.ble.android.mock.MockAndroidEnvironment
 import no.nordicsemi.kotlin.ble.client.android.CentralManager
 import no.nordicsemi.kotlin.ble.client.android.mock.internal.MockCentralManagerImpl
 import no.nordicsemi.kotlin.ble.client.mock.SimulationProvider
+import no.nordicsemi.kotlin.ble.environment.android.mock.LatestApi
+import no.nordicsemi.kotlin.ble.environment.android.mock.MockAndroidEnvironment
 
 /**
  * Creates a mock implementation of a [CentralManager] that can emulate scanning and connecting
