+++ environment-android-mock/src/main/AndroidManifest.xml
