+++ environment-android/build.gradle.kts
@@ -8,7 +8,7 @@ group = "no.nordicsemi.kotlin.ble"
 
 nordicNexusPublishing {
     POM_ARTIFACT_ID = "environment-android"
-    POM_NAME = "Android Environment Module"
+    POM_NAME = "Native Android Environment Module"
     POM_DESCRIPTION = "A part of Kotlin BLE Library providing Android-specific environment implementation."
     POM_URL = "https://github.com/NordicSemiconductor/Kotlin-BLE-Library"
     POM_SCM_URL = "https://github.com/NordicSemiconductor/Kotlin-BLE-Library"
